requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
# FILE 24: requirements.txt
# Python dependencies for AI integration scripts
# Uses OpenRouter (requests) — NOT anthropic SDK
# Install: pip install -r requirements.txt
# ━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━━
requests==2.31.0
ansible==9.2.0
pyyaml==6.0.1
tabulate==0.9.0
